Add ignore for environment files
